refactor: pass linter

+++ src/auth/auth.controller.ts
@@ -1,4 +1,12 @@
-import { BadRequestException, Body, Controller, Get, Logger, Post, Req, Res, UnauthorizedException } from '@nestjs/common';
+import {
+    Body,
+    Controller,
+    Logger,
+    Post,
+    Req,
+    Res,
+    UnauthorizedException,
+} from '@nestjs/common';
 import { BaseController } from '../common/base/base.controller';
 import { LoginDto, Role } from './types';
 import type { Request, Response } from 'express';
@@ -6,18 +14,17 @@ import { AuthService } from './auth.service';
 import { CookieService } from '../common/utils/cookie/cookie.service';
 import { BaseResponse } from '../common/base/base.response';
 import { Public, Roles } from './auth.decorator';
-import { JwtService } from '@nestjs/jwt';
-import { TypedConfigService } from '../common/typed-config/typed-config.service';
+import 'cookie-parser';
 
 @Controller('auth')
 export class AuthController extends BaseController {
     constructor(
         private service: AuthService,
         private cookieService: CookieService,
-        private jwtService: JwtService,
-        private config: TypedConfigService
-    ) { super() }
-    
+    ) {
+        super();
+    }
+
     @Public()
     @Roles(Role.Unauthenticated)
     @Post('login')
@@ -25,40 +32,45 @@ export class AuthController extends BaseController {
         @Body() dto: LoginDto,
         @Res({ passthrough: true }) res: Response,
     ) {
-        const { refreshPayload, jwtPayload, user } = await this.service.login(dto);
-        
+        const { refreshPayload, jwtPayload, user } =
+            await this.service.login(dto);
+
         this.cookieService.createJwt(res, jwtPayload);
         this.cookieService.createRefresh(res, refreshPayload);
         this.cookieService.createDummy(res);
 
-        Logger.log({jwtPayload})
-        return new BaseResponse({user});
+        Logger.log({ jwtPayload });
+        return new BaseResponse({ user });
     }
-    
+
     @Public()
     @Roles(Role.Unauthenticated)
     @Post('refresh')
     async refresh(
         @Req() req: Request,
-        @Res({passthrough: true}) res: Response
+        @Res({ passthrough: true }) res: Response,
     ) {
-        Logger.log('REFRESH')
-        const oldRefreshPayload = req.signedCookies['refresh']
-        
-        Logger.log('BEFORE TRY')
+        const oldRefreshPayload = req.signedCookies['refresh'] as
+            | string
+            | undefined;
+
+        if (!oldRefreshPayload) throw new Error();
+
         try {
-            const { refreshId } = JSON.parse(oldRefreshPayload)
+            const { refreshId } = JSON.parse(oldRefreshPayload) as {
+                refreshId: string;
+            };
 
-            Logger.log({refreshId})
-            if (!refreshId) throw new Error()
+            if (!refreshId) throw new Error();
 
-            const {refreshPayload, jwtPayload} = await this.service.refresh(refreshId)
-            this.cookieService.createRefresh(res, refreshPayload)
-            this.cookieService.createJwt(res, jwtPayload)
-            this.cookieService.createDummy(res)
+            const { refreshPayload, jwtPayload } =
+                await this.service.refresh(refreshId);
+            this.cookieService.createRefresh(res, refreshPayload);
+            this.cookieService.createJwt(res, jwtPayload);
+            this.cookieService.createDummy(res);
         } catch (err) {
-            Logger.error(err)
-            throw new UnauthorizedException('Please log in again')
+            Logger.error(err);
+            throw new UnauthorizedException('Please log in again');
         }
 
         return new BaseResponse();
@@ -70,15 +82,22 @@ export class AuthController extends BaseController {
         @Res({ passthrough: true }) res: Response,
         @Req() req: Request,
     ) {
-        const refreshPayload = req.signedCookies['refresh'];
-        Logger.log({refreshPayload})
-        try {
-            const { refreshId } = JSON.parse(refreshPayload)
-            if (!refreshId) throw new Error()
+        const refreshPayload = req.signedCookies['refresh'] as
+            | string
+            | undefined;
 
-            await this.service.logout(refreshId)
+        if (!refreshPayload) throw new Error();
+
+        try {
+            const { refreshId } = JSON.parse(refreshPayload) as {
+                refreshId: string;
+            };
+
+            if (!refreshId) throw new Error();
+
+            await this.service.logout(refreshId);
         } catch (err) {
-            Logger.warn('refreshPayload does not have valid content');
+            Logger.warn('refreshPayload does not have valid content', err);
         }
 
         this.cookieService.removeJwt(res);
